Adding roslint to vector_map and applying suggestions.

See merge request autowarefoundation/autoware.ai/common!80

+++ vector_map/include/vector_map/vector_map.h
@@ -17,7 +17,6 @@
 #ifndef VECTOR_MAP_VECTOR_MAP_H
 #define VECTOR_MAP_VECTOR_MAP_H
 
-#include <fstream>
 #include <ros/ros.h>
 #include <geometry_msgs/Point.h>
 #include <geometry_msgs/Quaternion.h>
@@ -55,6 +54,11 @@
 #include <vector_map_msgs/WallArray.h>
 #include <vector_map_msgs/FenceArray.h>
 #include <vector_map_msgs/RailCrossingArray.h>
+
+#include <fstream>
+#include <map>
+#include <string>
+#include <vector>
 
 namespace vector_map
 {
@@ -124,7 +128,7 @@ using vector_map_msgs::WallArray;
 using vector_map_msgs::FenceArray;
 using vector_map_msgs::RailCrossingArray;
 
-using category_t = unsigned long long;
+using category_t = uint32_t;
 
 enum Category : category_t
 {
@@ -297,7 +301,7 @@ std::vector<T> parse(const std::string& csv_file)
 {
   std::ifstream ifs(csv_file.c_str());
   std::string line;
-  std::getline(ifs, line); // remove first line
+  std::getline(ifs, line);  // remove first line
   std::vector<T> objs;
   while (std::getline(ifs, line))
   {
@@ -308,42 +312,6 @@ std::vector<T> parse(const std::string& csv_file)
   }
   return objs;
 }
-
-/* namespace */
-/* { */
-/* void updatePoint(std::map<Key<Point>, Point>& map, const PointArray& msg); */
-/* void updateVector(std::map<Key<Vector>, Vector>& map, const VectorArray& msg); */
-/* void updateLine(std::map<Key<Line>, Line>& map, const LineArray& msg); */
-/* void updateArea(std::map<Key<Area>, Area>& map, const AreaArray& msg); */
-/* void updatePole(std::map<Key<Pole>, Pole>& map, const PoleArray& msg); */
-/* void updateBox(std::map<Key<Box>, Box>& map, const BoxArray& msg); */
-/* void updateDTLane(std::map<Key<DTLane>, DTLane>& map, const DTLaneArray& msg); */
-/* void updateNode(std::map<Key<Node>, Node>& map, const NodeArray& msg); */
-/* void updateLane(std::map<Key<Lane>, Lane>& map, const LaneArray& msg); */
-/* void updateWayArea(std::map<Key<WayArea>, WayArea>& map, const WayAreaArray& msg); */
-/* void updateRoadEdge(std::map<Key<RoadEdge>, RoadEdge>& map, const RoadEdgeArray& msg); */
-/* void updateGutter(std::map<Key<Gutter>, Gutter>& map, const GutterArray& msg); */
-/* void updateCurb(std::map<Key<Curb>, Curb>& map, const CurbArray& msg); */
-/* void updateWhiteLine(std::map<Key<WhiteLine>, WhiteLine>& map, const WhiteLineArray& msg); */
-/* void updateStopLine(std::map<Key<StopLine>, StopLine>& map, const StopLineArray& msg); */
-/* void updateZebraZone(std::map<Key<ZebraZone>, ZebraZone>& map, const ZebraZoneArray& msg); */
-/* void updateCrossWalk(std::map<Key<CrossWalk>, CrossWalk>& map, const CrossWalkArray& msg); */
-/* void updateRoadMark(std::map<Key<RoadMark>, RoadMark>& map, const RoadMarkArray& msg); */
-/* void updateRoadPole(std::map<Key<RoadPole>, RoadPole>& map, const RoadPoleArray& msg); */
-/* void updateRoadSign(std::map<Key<RoadSign>, RoadSign>& map, const RoadSignArray& msg); */
-/* void updateSignal(std::map<Key<Signal>, Signal>& map, const SignalArray& msg); */
-/* void updateStreetLight(std::map<Key<StreetLight>, StreetLight>& map, const StreetLightArray& msg); */
-/* void updateUtilityPole(std::map<Key<UtilityPole>, UtilityPole>& map, const UtilityPoleArray& msg); */
-/* void updateGuardRail(std::map<Key<GuardRail>, GuardRail>& map, const GuardRailArray& msg); */
-/* void updateSideWalk(std::map<Key<SideWalk>, SideWalk>& map, const SideWalkArray& msg); */
-/* void updateDriveOnPortion(std::map<Key<DriveOnPortion>, DriveOnPortion>& map, const DriveOnPortionArray& msg); */
-/* void updateCrossRoad(std::map<Key<CrossRoad>, CrossRoad>& map, const CrossRoadArray& msg); */
-/* void updateSideStrip(std::map<Key<SideStrip>, SideStrip>& map, const SideStripArray& msg); */
-/* void updateCurveMirror(std::map<Key<CurveMirror>, CurveMirror>& map, const CurveMirrorArray& msg); */
-/* void updateWall(std::map<Key<Wall>, Wall>& map, const WallArray& msg); */
-/* void updateFence(std::map<Key<Fence>, Fence>& map, const FenceArray& msg); */
-/* void updateRailCrossing(std::map<Key<RailCrossing>, RailCrossing>& map, const RailCrossingArray& msg); */
-/* } // namespace */
 
 class VectorMap
 {
@@ -530,7 +498,7 @@ geometry_msgs::Point convertPointToGeomPoint(const Point& point);
 Point convertGeomPointToPoint(const geometry_msgs::Point& geom_point);
 geometry_msgs::Quaternion convertVectorToGeomQuaternion(const Vector& vector);
 Vector convertGeomQuaternionToVector(const geometry_msgs::Quaternion& geom_quaternion);
-} // namespace vector_map
+}  // namespace vector_map
 
 std::ostream& operator<<(std::ostream& os, const vector_map::Point& obj);
 std::ostream& operator<<(std::ostream& os, const vector_map::Vector& obj);
@@ -598,4 +566,4 @@ std::istream& operator>>(std::istream& is, vector_map::Wall& obj);
 std::istream& operator>>(std::istream& is, vector_map::Fence& obj);
 std::istream& operator>>(std::istream& is, vector_map::RailCrossing& obj);
 
-#endif // VECTOR_MAP_VECTOR_MAP_H
+#endif  // VECTOR_MAP_VECTOR_MAP_H
